requirements.txt
PyJWT[crypto]==1.7.1
requests==2.22.0